🐛 Fix :  groupJoin  에러 수정

+++ backend/src/routers/groupJoinRouter.js
@@ -14,7 +14,6 @@ groupJoinRouter.post('/mygroups/:groupId', loginRequired, async (req, res, next)
     const userId = req.currentUserId;
     const groupId = req.params.groupId;
     const state = '대기';
-    console.log('req.body', req.body);
 
     //다른 그룹 중복 가입 방지
     const group = await groupJoinService.getUserGroup({ userId });
@@ -23,11 +22,8 @@ groupJoinRouter.post('/mygroups/:groupId', loginRequired, async (req, res, next)
       return;
     }
 
-    //유저스키마에 groupId 정보 업뎃
-    const updatedUser = await userAuthService.setUserGroup({ userId, groupId });
-
     const result = await groupJoinService.groupJoin({ groupId, userId, state });
-    res.json({ result, updatedUser, message: '등록 성공' });
+    res.json({ result, message: '등록 성공' });
     return;
   } catch (err) {
     next(err);
@@ -35,17 +31,16 @@ groupJoinRouter.post('/mygroups/:groupId', loginRequired, async (req, res, next)
 });
 
 //유저가 가입한 그룹 조회
-groupJoinRouter.get('/mygroups/:userId', loginRequired, async (req, res) => {
-  const userId = req.params.userId;
-  const result = await groupJoinService.getMyGroup({userId});
-  res.status(200).json({ result });
+groupJoinRouter.get('/mygroups', loginRequired, async (req, res) => {
+  const userId = req.currentUserId;
+  const result = await userAuthService.getMyGroup({ userId });
+  res.status(200).json(result);
   return;
 });
 
-// 그룹 가입 대기자 리스트 - 관리자용 - 완
+// 그룹 가입 대기자 리스트 - 관리자용
 groupJoinRouter.get('/mygroups/:groupId/waiting', loginRequired, async (req, res) => {
   const groupId = req.params.groupId;
-  console.log('groupId', groupId);
 
   const result = await groupJoinService.getWaitingList({ groupId });
   console.log('result', result);
@@ -53,15 +48,18 @@ groupJoinRouter.get('/mygroups/:groupId/waiting', loginRequired, async (req, res
   return;
 });
 
-/// 유저 가입 대기 -> 승인으로 관리자 승인에 의한 상태 변경 - 관리자용
-groupJoinRouter.put('/mygroups/:groupId/:userId/approval', loginRequired,  async (req, res) => {
+// 유저 가입 대기 -> 승인으로 관리자 승인에 의한 상태 변경 - 관리자용
+groupJoinRouter.put('/mygroups/:groupId/:userId/approval', loginRequired, async (req, res) => {
   const groupId = req.params.groupId;
   const userId = req.params.userId;
+
+  //유저스키마에 groupId 정보 업뎃
+  const updatedUser = await userAuthService.setUserGroup({ userId, groupId });
 
   const result = await groupJoinService.setJoinedGroup({ groupId, userId });
 
   if (result) {
-    res.status(200).json({ result, message: '가입승인' });
+    res.status(200).json({ result, updatedUser, message: '가입승인' });
   } else {
     res.status(400).json({ message: '대기 중인 대기자를 찾을 수 없습니다.' });
   }
@@ -69,16 +67,14 @@ groupJoinRouter.put('/mygroups/:groupId/:userId/approval', loginRequired,  async
 });
 
 //유저 가입 대기 -> 거절로 관리자 거절에 의한 상태 변경
-groupJoinRouter.delete('/mygroups/:groupId/:userId/rejection', loginRequired,  async (req, res) => {
+groupJoinRouter.delete('/mygroups/:groupId/:userId/rejection', loginRequired, async (req, res) => {
   const groupId = req.params.groupId;
   const userId = req.params.userId;
 
   //유저스키마에 groupId 정보 삭제
-  const updatedUser = await userAuthService.deleteGroupId({ groupId });
-  console.log('groupId업데이트 된 유저: ', updatedUser);
+  const updatedUser = await userAuthService.deleteGroupId({ groupId, userId });
 
   const result = await groupJoinService.deletedGroupJoinByOwner({ groupId, userId });
-  console.log('result:', result);
 
   if (result) {
     res.status(200).json({ result, updatedUser });
@@ -89,17 +85,15 @@ groupJoinRouter.delete('/mygroups/:groupId/:userId/rejection', loginRequired,  a
 });
 
 // 유저가 가입한 그룹 탈퇴 - 완
-groupJoinRouter.delete('/mygroups/:groupId/:userId', loginRequired, async (req, res) => {
+groupJoinRouter.delete('/mygroups/:groupId', loginRequired, async (req, res) => {
   const groupId = req.params.groupId;
   const userId = req.currentUserId;
 
   //유저스키마에 groupId 정보 삭제
-  const toUpdate = { groupId };
   const updatedUser = await userAuthService.deleteGroupId({ groupId, userId });
   console.log('groupId업데이트 된 유저: ', updatedUser);
 
   const result = await groupJoinService.deleteMyGroup({ userId });
-  console.log('1', result);
   res.status(200).json({ result, updatedUser });
   return;
 });
